reorganization and bug fix
