+++ py4DSTEM/process/latticevectors/__init__.py
@@ -1,4 +1,3 @@
 from py4DSTEM.process.latticevectors.initialguess import *
 from py4DSTEM.process.latticevectors.index import *
 from py4DSTEM.process.latticevectors.fit import *
-from py4DSTEM.process.latticevectors.strain import *
