+++ py4DSTEM/process/latticevectors/index.py
@@ -34,61 +34,6 @@ def get_selected_lattice_vectors(gx, gy, i0, i1, i2):
     return (g1x, g1y), (g2x, g2y)
 
 
-def index_bragg_directions(x0, y0, gx, gy, g1, g2):
-    """
-    From an origin (x0,y0), a set of reciprocal lattice vectors gx,gy, and an pair of
-    lattice vectors g1=(g1x,g1y), g2=(g2x,g2y), find the indices (h,k) of all the
-    reciprocal lattice directions.
-
-    The approach is to solve the matrix equation
-            ``alpha = beta * M``
-    where alpha is the 2xN array of the (x,y) coordinates of N measured bragg directions,
-    beta is the 2x2 array of the two lattice vectors u,v, and M is the 2xN array of the
-    h,k indices.
-
-    Args:
-        x0 (float): x-coord of origin
-        y0 (float): y-coord of origin
-        gx (1d array): x-coord of the reciprocal lattice vectors
-        gy (1d array): y-coord of the reciprocal lattice vectors
-        g1 (2-tuple of floats): g1x,g1y
-        g2 (2-tuple of floats): g2x,g2y
-
-    Returns:
-        (3-tuple) A 3-tuple containing:
-
-            * **h**: *(ndarray of ints)* first index of the bragg directions
-            * **k**: *(ndarray of ints)* second index of the bragg directions
-            * **bragg_directions**: *(PointList)* a 4-coordinate PointList with the
-              indexed bragg directions; coords 'qx' and 'qy' contain bragg_x and bragg_y
-              coords 'h' and 'k' contain h and k.
-    """
-    # Get beta, the matrix of lattice vectors
-    beta = np.array([[g1[0], g2[0]], [g1[1], g2[1]]])
-
-    # Get alpha, the matrix of measured bragg angles
-    alpha = np.vstack([gx - x0, gy - y0])
-
-    # Calculate M, the matrix of peak positions
-    M = lstsq(beta, alpha, rcond=None)[0].T
-    M = np.round(M).astype(int)
-
-    # Get h,k
-    h = M[:, 0]
-    k = M[:, 1]
-
-    # Store in a PointList
-    coords = [("qx", float), ("qy", float), ("h", int), ("k", int)]
-    temp_array = np.zeros([], dtype=coords)
-    bragg_directions = PointList(data=temp_array)
-    bragg_directions.add_data_by_field((gx, gy, h, k))
-    mask = np.zeros(bragg_directions["qx"].shape[0])
-    mask[0] = 1
-    bragg_directions.remove(mask)
-
-    return h, k, bragg_directions
-
-
 def generate_lattice(ux, uy, vx, vy, x0, y0, Q_Nx, Q_Ny, h_max=None, k_max=None):
     """
     Returns a full reciprocal lattice stretching to the limits of the diffraction pattern
@@ -163,82 +108,6 @@ def generate_lattice(ux, uy, vx, vy, x0, y0, Q_Nx, Q_Ny, h_max=None, k_max=None)
     return ideal_lattice
 
 
-def add_indices_to_braggvectors(
-    braggpeaks, lattice, maxPeakSpacing, qx_shift=0, qy_shift=0, mask=None
-):
-    """
-    Using the peak positions (qx,qy) and indices (h,k) in the PointList lattice,
-    identify the indices for each peak in the PointListArray braggpeaks.
-    Return a new braggpeaks_indexed PointListArray, containing a copy of braggpeaks plus
-    three additional data columns -- 'h','k', and 'index_mask' -- specifying the peak
-    indices with the ints (h,k) and indicating whether the peak was successfully indexed
-    or not with the bool index_mask. If `mask` is specified, only the locations where
-    mask is True are indexed.
-
-    Args:
-        braggpeaks (PointListArray): the braggpeaks to index. Must contain
-            the coordinates 'qx', 'qy', and 'intensity'
-        lattice (PointList): the positions (qx,qy) of the (h,k) lattice points.
-            Must contain the coordinates 'qx', 'qy', 'h', and 'k'
-        maxPeakSpacing (float): Maximum distance from the ideal lattice points
-            to include a peak for indexing
-        qx_shift,qy_shift (number): the shift of the origin in the `lattice` PointList
-            relative to the `braggpeaks` PointListArray
-        mask (bool): Boolean mask, same shape as the pointlistarray, indicating which
-            locations should be indexed. This can be used to index different regions of
-            the scan with different lattices
-
-    Returns:
-        (PointListArray): The original braggpeaks pointlistarray, with new coordinates
-        'h', 'k', containing the indices of each indexable peak.
-    """
-
-    # assert isinstance(braggpeaks,BraggVectors)
-    # assert isinstance(lattice, PointList)
-    # assert np.all([name in lattice.dtype.names for name in ('qx','qy','h','k')])
-
-    if mask is None:
-        mask = np.ones(braggpeaks.Rshape, dtype=bool)
-
-    assert (
-        mask.shape == braggpeaks.Rshape
-    ), "mask must have same shape as pointlistarray"
-    assert mask.dtype == bool, "mask must be boolean"
-
-    coords = [
-        ("qx", float),
-        ("qy", float),
-        ("intensity", float),
-        ("h", int),
-        ("k", int),
-    ]
-
-    indexed_braggpeaks = PointListArray(
-        dtype=coords,
-        shape=braggpeaks.Rshape,
-    )
-
-    # loop over all the scan positions
-    for Rx, Ry in tqdmnd(mask.shape[0], mask.shape[1]):
-        if mask[Rx, Ry]:
-            pl = braggpeaks.cal[Rx, Ry]
-            for i in range(pl.data.shape[0]):
-                r2 = (pl.data["qx"][i] - lattice.data["qx"] + qx_shift) ** 2 + (
-                    pl.data["qy"][i] - lattice.data["qy"] + qy_shift
-                ) ** 2
-                ind = np.argmin(r2)
-                if r2[ind] <= maxPeakSpacing**2:
-                    indexed_braggpeaks[Rx, Ry].add_data_by_field(
-                        (
-                            pl.data["qx"][i],
-                            pl.data["qy"][i],
-                            pl.data["intensity"][i],
-                            lattice.data["h"][ind],
-                            lattice.data["k"][ind],
-                        )
-                    )
-
-    return indexed_braggpeaks
 
 
 def bragg_vector_intensity_map_by_index(braggpeaks, h, k, symmetric=False):
