+++ py4DSTEM/process/strain/__init__.py
@@ -1,0 +1,2 @@
+from py4DSTEM.process.strain.strain import StrainMap
+from py4DSTEM.process.strain.latticevectors import *
